requirements.txt
pillow
tk
pywin32
psutil